Relocate resolver controller helper into the plugin shared bundle
- move createResolverController from src/shared/src/utils/resolver-controller.js to src/plugin/src/shared/resolver-controller.js because only plugin option plumbing uses it, keeping the shared bundle focused on cross-environment utilities
- update plugin modules/tests to import the helper from its new location and drop the shared export
