+++ src/plugin/src/shared/index.js
@@ -1,1 +1,2 @@
 export * from "@prettier-plugin-gml/shared";
+export * from "./resolver-controller.js";
