+++ src/plugin/test/resolver-controller.test.js
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
-import { createResolverController } from "../src/utils/resolver-controller.js";
+import { createResolverController } from "../src/shared/resolver-controller.js";
 
 describe("createResolverController", () => {
     it("returns defaults when no resolver is installed", () => {
