+++ src/shared/src/utils/index.js
@@ -11,6 +11,5 @@ export * from "./line-breaks.js";
 export * from "./number.js";
 export * from "./numeric-options.js";
 export * from "./object.js";
-export * from "./resolver-controller.js";
 export * from "./regexp.js";
 export * from "./string.js";
